% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercept0.R
\name{summary.force_intercept_model}
\alias{summary.force_intercept_model}
\title{Custom Summary for force_intercept_model}
\usage{
\method{summary}{force_intercept_model}(object, ...)
}
\arguments{
\item{object}{An object of class \code{force_intercept_model} (inherits from \code{lm}).}

\item{...}{Additional arguments passed to the \code{summary.lm} method.}
}
\value{
The summary object of class \code{summary.lm} and prints the corrected R² value.
}
\description{
Custom summary method for objects of class \code{force_intercept_model}. It prints
the regular linear model summary, along with the corrected R² value.
}
\examples{
x=c(19.5,40.8,45.2,55.3,24.9,49,50,29.7,32.3,42.6,54.2,52.9,64.3,53.6,34.1,18.1,66.7,57.2,37.1,58.1,67.2,53.5,63,42.1,34.9,71.3,23)
y=c(20.41,37.43,36.57,50.34,23.07,44.98,44.35,29.27,28.10,40.89,52.16,49.04,61.57,52.3,40.86,23.18,60.13,51.02,35.37,54.46,59.88,45.96,53.9,39.33,34.58,61.35,25.37)
df= data.frame (x,y)

if(!require(remotes)) install.packages("remotes")
if (!requireNamespace("intercept0", quietly = TRUE)) {
     remotes::install_github("agronomy4future/intercept0")
}
library(remotes)
library(intercept0)

model= intercept0(y~x,data=df)
summary(model)
}
